requirements.txt
streamlit>=1.20.0
PyPDF2>=3.0.0
python-docx>=0.8.11
python-dotenv>=0.19.0
google-generativeai>=0.3.0
reportlab>=3.6.0
Pillow>=9.0.0 